+++ ReleaseNotes.md
@@ -2,6 +2,19 @@ Versity ScoutFS Release Notes
 =============================
 
 ---
+v1.x
+\
+*TBD*
+
+
+* **Add scoutfs(1) change-quorum-config command**
+\
+  Add a change-quorum-config command to scoutfs(1) to change the quorum
+  configuration stored in the metadata device while the file system is
+  unmounted.   This can be used to change the mounts that will
+  participate in quorum and the IP addresses they use.
+
+---
 v1.0
 \
 *Nov 8, 2021*